chore(dev-console): move AttributeEditor to new component
